pinterest_automation/requirements.txt
# HTTP requests
requests>=2.28.1

# Environment variables
python-dotenv>=0.21.0

# Data processing
pandas>=1.5.0
numpy>=1.23.0

# Logging
colorlog>=6.7.0

# Scheduling
apscheduler>=3.9.1

# Database (for storing pins and analytics)
sqlalchemy>=1.4.40

# Image processing (optional, for image optimization)
pillow>=9.2.0

# For testing
pytest>=7.0.0
